install.sh, npm is only invoked if it exists

+++ install.sh
@@ -8,8 +8,10 @@ pip install pip wheel --upgrade
 pip install -r requirements.txt
 NEW_RELIC_EXTENSIONS=false pip install --no-binary :all: newrelic==2.82.0.62
 
-# clean install node.js deps
-if [ -e package-lock.json ]; then
-    rm -rf node_modules
-    npm install
+if which npm; then
+    # clean install node.js deps
+    if [ -e package-lock.json ]; then
+        rm -rf node_modules
+        npm install
+    fi
 fi
